refactor(db): split helpers/enums.ts by domain
Bean: db-pgjf
