+++ packages/db/src/helpers/enums/api.ts
@@ -1,0 +1,246 @@
+/**
+ * API key and audit const arrays for varchar CHECK constraints.
+ * Values sourced from @pluralscape/types union types.
+ */
+
+import { ALL_API_KEY_SCOPES, type ApiKey, type AuditEventType } from "@pluralscape/types";
+
+/**
+ * Naming convention: TABLE_COLUMN (e.g. API_KEY_KEY_TYPES = api_keys.key_type).
+ * Each array name maps to the table and column it constrains, so
+ * `API_KEY_KEY_TYPES` → `api_keys.key_type`, `API_KEY_SCOPES` → `api_keys.scopes`, etc.
+ */
+export const API_KEY_KEY_TYPES = [
+  "metadata",
+  "crypto",
+] as const satisfies readonly ApiKey["keyType"][];
+
+export { ALL_API_KEY_SCOPES as API_KEY_SCOPES };
+
+export const AUDIT_EVENT_TYPES = [
+  "auth.register",
+  "auth.login",
+  "auth.login-failed",
+  "auth.logout",
+  "auth.password-changed",
+  "auth.recovery-key-used",
+  "auth.key-created",
+  "auth.key-revoked",
+  "data.export",
+  "data.import",
+  "data.purge",
+  "api-key.created",
+  "api-key.revoked",
+  "settings.changed",
+  "member.created",
+  "member.archived",
+  "member.deleted",
+  "sharing.granted",
+  "sharing.revoked",
+  "bucket.key_rotation.initiated",
+  "bucket.key_rotation.chunk_completed",
+  "bucket.key_rotation.completed",
+  "bucket.key_rotation.failed",
+  "bucket.key_rotation.retried",
+  "device.security.jailbreak_warning_shown",
+  "auth.password-reset-via-recovery",
+  "auth.recovery-key-regenerated",
+  "auth.device-transfer-initiated",
+  "auth.device-transfer-completed",
+  "auth.device-transfer-approved",
+  "auth.email-changed",
+  "system.created",
+  "system.profile-updated",
+  "system.deleted",
+  "system.purged",
+  "system.duplicated",
+  "snapshot.created",
+  "snapshot.deleted",
+  "group.created",
+  "group.updated",
+  "group.archived",
+  "group.restored",
+  "group.moved",
+  "group-membership.added",
+  "group-membership.removed",
+  "custom-front.created",
+  "custom-front.updated",
+  "custom-front.archived",
+  "custom-front.restored",
+  "group.deleted",
+  "custom-front.deleted",
+  "auth.biometric-enrolled",
+  "auth.biometric-verified",
+  "auth.biometric-failed",
+  "settings.pin-set",
+  "settings.pin-removed",
+  "settings.pin-verified",
+  "settings.nomenclature-updated",
+  "setup.step-completed",
+  "setup.completed",
+  "member.updated",
+  "member.duplicated",
+  "member.restored",
+  "member-photo.created",
+  "member-photo.archived",
+  "member-photo.restored",
+  "member-photo.deleted",
+  "member-photo.reordered",
+  "field-definition.created",
+  "field-definition.updated",
+  "field-definition.archived",
+  "field-definition.restored",
+  "field-definition.deleted",
+  "field-value.set",
+  "field-value.updated",
+  "field-value.deleted",
+  "structure-entity-type.created",
+  "structure-entity-type.updated",
+  "structure-entity-type.archived",
+  "structure-entity-type.restored",
+  "structure-entity-type.deleted",
+  "structure-entity.created",
+  "structure-entity.updated",
+  "structure-entity.archived",
+  "structure-entity.restored",
+  "structure-entity.deleted",
+  "structure-entity-link.created",
+  "structure-entity-link.deleted",
+  "structure-entity-member-link.added",
+  "structure-entity-member-link.removed",
+  "structure-entity-association.created",
+  "structure-entity-association.deleted",
+  "relationship.created",
+  "relationship.updated",
+  "relationship.archived",
+  "relationship.restored",
+  "relationship.deleted",
+  "lifecycle-event.created",
+  "lifecycle-event.archived",
+  "lifecycle-event.restored",
+  "lifecycle-event.deleted",
+  "lifecycle-event.updated",
+  "timer-config.created",
+  "timer-config.updated",
+  "timer-config.archived",
+  "timer-config.restored",
+  "timer-config.deleted",
+  "check-in-record.created",
+  "check-in-record.responded",
+  "check-in-record.dismissed",
+  "check-in-record.archived",
+  "check-in-record.deleted",
+  "check-in-record.restored",
+  "innerworld-region.created",
+  "innerworld-region.updated",
+  "innerworld-region.archived",
+  "innerworld-region.restored",
+  "innerworld-region.deleted",
+  "innerworld-entity.created",
+  "innerworld-entity.updated",
+  "innerworld-entity.archived",
+  "innerworld-entity.restored",
+  "innerworld-entity.deleted",
+  "innerworld-canvas.created",
+  "innerworld-canvas.updated",
+  "blob.upload-requested",
+  "blob.confirmed",
+  "blob.archived",
+  "fronting-report.created",
+  "fronting-report.updated",
+  "fronting-report.archived",
+  "fronting-report.restored",
+  "fronting-report.deleted",
+  "fronting-session.created",
+  "fronting-session.updated",
+  "fronting-session.ended",
+  "fronting-session.archived",
+  "fronting-session.restored",
+  "fronting-session.deleted",
+  "fronting-comment.created",
+  "fronting-comment.updated",
+  "fronting-comment.archived",
+  "fronting-comment.restored",
+  "fronting-comment.deleted",
+  "webhook-config.created",
+  "webhook-config.updated",
+  "webhook-config.archived",
+  "webhook-config.restored",
+  "webhook-config.deleted",
+  "webhook-config.secret-rotated",
+  "webhook-delivery.deleted",
+  "channel.created",
+  "channel.updated",
+  "channel.archived",
+  "channel.restored",
+  "channel.deleted",
+  "message.created",
+  "message.updated",
+  "message.archived",
+  "message.restored",
+  "message.deleted",
+  "board-message.created",
+  "board-message.updated",
+  "board-message.pinned",
+  "board-message.unpinned",
+  "board-message.reordered",
+  "board-message.archived",
+  "board-message.restored",
+  "board-message.deleted",
+  "note.created",
+  "note.updated",
+  "note.archived",
+  "note.restored",
+  "note.deleted",
+  "poll.created",
+  "poll.updated",
+  "poll.closed",
+  "poll.archived",
+  "poll.restored",
+  "poll.deleted",
+  "poll-vote.cast",
+  "poll-vote.vetoed",
+  "poll-vote.updated",
+  "poll-vote.archived",
+  "acknowledgement.created",
+  "acknowledgement.confirmed",
+  "acknowledgement.archived",
+  "acknowledgement.restored",
+  "acknowledgement.deleted",
+  // ── Privacy: buckets ──
+  "bucket.created",
+  "bucket.updated",
+  "bucket.archived",
+  "bucket.restored",
+  "bucket.deleted",
+  "bucket-content-tag.tagged",
+  "bucket-content-tag.untagged",
+  "field-bucket-visibility.set",
+  "field-bucket-visibility.removed",
+  // ── Privacy: friends ──
+  "friend-code.generated",
+  "friend-code.redeemed",
+  "friend-code.archived",
+  "friend-connection.created",
+  "friend-connection.accepted",
+  "friend-connection.rejected",
+  "friend-connection.blocked",
+  "friend-connection.removed",
+  "friend-connection.archived",
+  "friend-connection.restored",
+  "friend-visibility.updated",
+  "friend-bucket-assignment.assigned",
+  "friend-bucket-assignment.unassigned",
+  // ── Notifications ──
+  "device-token.registered",
+  "device-token.revoked",
+  "device-token.updated",
+  "device-token.deleted",
+  "notification-config.updated",
+  "friend-notification-preference.updated",
+  // ── Import ──
+  "import-job.created",
+  "import-job.updated",
+  "import-job.completed",
+  "import-job.failed",
+] as const satisfies readonly AuditEventType[];
